Update for Exchange 2016 CU 15 ActivationPreference
Hello. I have upgraded to Exchange 2016 CU 15 and the Activation Preference is no longer showing.

The line in the script that gets the Activation Preference i is not getting the preference.

$pref = ($database | Select-Object -ExpandProperty ActivationPreference | Where-Object {$_.Key -ieq $mailboxserver}).Value

If I run the script using an Exchange 2010 shell is works but not when it is Exchange 2016. Can someone help me fix the script so I get the preference again.